% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{as_expression_matrix}
\alias{as_expression_matrix}
\title{Coerce input to expression matrix}
\usage{
as_expression_matrix(x)
}
\arguments{
\item{x}{Numeric matrix-like object with samples in rows and genes in columns.}
}
\value{
Numeric matrix.
}
\description{
Coerce input to expression matrix
}
