% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_studies.R
\name{available_case_studies}
\alias{available_case_studies}
\title{List bundled case-study ids}
\usage{
available_case_studies()
}
\value{
Character vector of case-study dataset ids.
}
\description{
List bundled case-study ids
}
