% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_studies.R
\name{case_study_table}
\alias{case_study_table}
\title{Read case-study output table}
\usage{
case_study_table(dataset = "tcga_blca", table_name = "phenotype_associations")
}
\arguments{
\item{dataset}{Dataset id. One of \code{available_case_studies()}.}

\item{table_name}{Output table basename without extension.}
}
\value{
A data frame.
}
\description{
Read case-study output table
}
