% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{classification_err}
\alias{classification_err}
\title{Classification error of a vector}
\usage{
classification_err(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Classification error in the range \code{0} to \code{1}.
}
\description{
Classification error of a vector
}
