% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{compute_dissimilarity}
\alias{compute_dissimilarity}
\title{Compute dissimilarity matrix}
\usage{
compute_dissimilarity(adjacency)
}
\arguments{
\item{adjacency}{Square adjacency matrix.}
}
\value{
Gene-by-gene dissimilarity matrix.
}
\description{
Compute dissimilarity matrix
}
