% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_decomposition.R
\name{core_decompose_weighted}
\alias{core_decompose_weighted}
\title{Weighted core decomposition}
\usage{
core_decompose_weighted(x)
}
\arguments{
\item{x}{Weighted adjacency-like square matrix.}
}
\value{
List with integer \code{order} and numeric \code{degeneracy}.
}
\description{
Weighted core decomposition
}
