% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_decomposition.R
\name{erode_to_core}
\alias{erode_to_core}
\title{Erode clusters to core genes}
\usage{
erode_to_core(clusters, dissimilarity)
}
\arguments{
\item{clusters}{Integer cluster labels (\code{-1} for noise).}

\item{dissimilarity}{Gene-gene dissimilarity matrix.}
}
\value{
Updated core cluster labels.
}
\description{
Erode clusters to core genes
}
