% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedWGCNA-package.R
\docType{package}
\name{NestedWGCNA-package}
\alias{NestedWGCNA}
\alias{NestedWGCNA-package}
\title{NestedWGCNA}
\description{
Minimal two-stage gene co-expression analysis in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dai540/NestedWGCNA}
  \item Report bugs at \url{https://github.com/dai540/NestedWGCNA/issues}
}

}
\keyword{internal}
