% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_nested_wgcna}
\alias{run_nested_wgcna}
\title{Run nested two-stage workflow}
\usage{
run_nested_wgcna(
  x,
  mode = c("paper", "python_compat"),
  min_cgm_size = 100L,
  min_fgm_size = 20L,
  top_n_genes = 2000L,
  corr_method = c("pearson", "spearman")
)
}
\arguments{
\item{x}{Numeric matrix-like object (\verb{samples x genes}).}

\item{mode}{Similarity mode.}

\item{min_cgm_size}{Minimum CGM size.}

\item{min_fgm_size}{Minimum FGM size.}

\item{top_n_genes}{Number of genes used in stage 1.}

\item{corr_method}{Correlation method.}
}
\value{
Object of class \code{NestedWGCNAResult}.
}
\description{
Run nested two-stage workflow
}
