% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.NestedWGCNAResult}
\alias{summary.NestedWGCNAResult}
\title{Summarize a NestedWGCNA result}
\usage{
\method{summary}{NestedWGCNAResult}(object, ...)
}
\arguments{
\item{object}{A \code{NestedWGCNAResult}.}

\item{...}{Unused.}
}
\value{
A named list summary.
}
\description{
Summarize a NestedWGCNA result
}
