% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{top_variable_genes}
\alias{top_variable_genes}
\title{Select top variable genes}
\usage{
top_variable_genes(x, n = 1200)
}
\arguments{
\item{x}{Expression matrix (\verb{samples x genes}).}

\item{n}{Number of genes to keep.}
}
\value{
Matrix with top variable genes.
}
\description{
Select top variable genes
}
