% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{as_expression_matrix}
\alias{as_expression_matrix}
\title{Coerce input to numeric expression matrix}
\usage{
as_expression_matrix(x, require_names = TRUE)
}
\arguments{
\item{x}{Matrix-like object (\code{matrix} or \code{data.frame}) with rows as samples
and columns as genes.}

\item{require_names}{If \code{TRUE}, requires row and column names.}
}
\value{
Numeric matrix.
}
\description{
Coerce input to numeric expression matrix
}
