% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{compute_dissimilarity}
\alias{compute_dissimilarity}
\title{Compute dissimilarity matrix}
\usage{
compute_dissimilarity(
  adjacency,
  method = c("paper_metric", "python_compat", "tom", "mot", "custom"),
  custom_fun = NULL
)
}
\arguments{
\item{adjacency}{Adjacency matrix.}

\item{method}{One of \code{"paper_metric"}, \code{"python_compat"}, \code{"tom"}, \code{"mot"},
or \code{"custom"}.}

\item{custom_fun}{Optional function for custom dissimilarity.}
}
\value{
Gene-gene dissimilarity matrix.
}
\description{
Compute dissimilarity matrix
}
