% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{module_score}
\alias{module_score}
\title{Calculate per-sample module scores}
\usage{
module_score(x, assignment, prefix = "M_")
}
\arguments{
\item{x}{Expression matrix (\verb{samples x genes}).}

\item{assignment}{Integer module assignment vector for genes (\code{-1} = noise).}

\item{prefix}{Prefix for score column names.}
}
\value{
Data frame (\verb{samples x modules}) of module scores.
}
\description{
Calculate per-sample module scores
}
