% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedWGCNA-package.R
\docType{package}
\name{NestedWGCNA-package}
\alias{NestedWGCNA-package}
\alias{NestedWGCNA}
\title{NestedWGCNA}
\description{
Two-stage gene co-expression network analysis with coarse-grained and
fine-grained module discovery.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dai540.github.io/NestedWGCNA}
  \item \url{https://github.com/dai540/NestedWGCNA}
  \item Report bugs at \url{https://github.com/dai540/NestedWGCNA/issues}
}

}
\author{
Dai
}
