% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.NestedWGCNAResult}
\alias{summary.NestedWGCNAResult}
\title{Summarize nested analysis result}
\usage{
\method{summary}{NestedWGCNAResult}(object, ...)
}
\arguments{
\item{object}{\code{NestedWGCNAResult} object.}

\item{...}{Unused.}
}
\value{
A summary list.
}
\description{
Summarize nested analysis result
}
