% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{top_variable_genes}
\alias{top_variable_genes}
\title{Select top variable genes}
\usage{
top_variable_genes(x, n = 2000L)
}
\arguments{
\item{x}{Numeric matrix-like object.}

\item{n}{Number of genes to keep.}
}
\value{
Matrix with selected genes.
}
\description{
Select top variable genes
}
